Hackathon_Backend/requirements.txt
# MindVest Backend — Python Dependencies

# Web Framework
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Settings & Validation
pydantic>=2.6.4
pydantic-settings>=2.2.1

# Database
sqlalchemy>=2.0.29
supabase>=2.4.2

# Auth & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Web3
web3>=6.15.1

# HTTP Client
httpx>=0.27.0

# Finance / ML
# yfinance 0.2.54+ dropped lru-dict (no more C compiler needed on Windows)
yfinance>=0.2.54
# Prophet requires pystan which needs C++ — install separately if needed:
# pip install prophet
# For now, use neuralprophet or skip prophet install

# AI / LLM
google-generativeai>=0.5.2
openai>=1.23.2

# Utilities
python-dotenv>=1.0.1
pandas>=2.2.1
